Initial commit
Copied my Eliza plugin from Jasper to Naomi.

This still needs a lot of work to really be useful. Right now
the state of voice control is not nearly accurate enough with
a completely open dictionary. My hope is to eventually get a
combination of PocketSphinx and Deepspeech working with this.